Avoid Roles&Permission seeders when it's not needed
