+++ tests/Feature/Http/Api/ApiHostsTest.php
@@ -2,20 +2,23 @@
 
 namespace Tests\Feature\Http\Api;
 
-use App\Enums\Roles;
 use App\Models\Host;
 use App\Models\Hostgroup;
 use App\Models\User;
 use Tests\ApiTestCase;
+use Tests\Traits\InteractsWithPermissions;
 
 class ApiHostsTest extends ApiTestCase
 {
+    use InteractsWithPermissions;
+
     private User $user;
 
     public function setUp(): void
     {
         parent::setUp();
-        $this->disableAuthorization();
+
+        $this->disablePermissionsCheck();
         $this->user = User::factory()->create();
     }
 
@@ -49,7 +52,7 @@ class ApiHostsTest extends ApiTestCase
             ],
             'relationships' => [
                 'groups' => [
-                    'data' => $groups->map(fn (Hostgroup $group) => [
+                    'data' => $groups->map(fn(Hostgroup $group) => [
                         'type' => 'hostgroups',
                         'id' => (string) $group->getRouteKey(),
                     ])->all(),
@@ -182,7 +185,7 @@ class ApiHostsTest extends ApiTestCase
             ],
             'relationships' => [
                 'groups' => [
-                    'data' => $newGroups->map(fn (Hostgroup $group) => [
+                    'data' => $newGroups->map(fn(Hostgroup $group) => [
                         'type' => 'hostgroups',
                         'id' => (string) $group->getRouteKey(),
                     ])->all(),
