+++ tests/Feature/Http/Controllers/HostgroupControllerTest.php
@@ -17,24 +17,26 @@
 
 namespace Tests\Feature\Http\Controllers;
 
-use App\Enums\Roles;
 use App\Models\Host;
 use App\Models\Hostgroup;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
+use Tests\Traits\InteractsWithPermissions;
 
 class HostgroupControllerTest extends TestCase
 {
     use RefreshDatabase;
+    use InteractsWithPermissions;
 
     private User $user;
 
     public function setUp(): void
     {
         parent::setUp();
+
+        $this->disablePermissionsCheck();
         $this->user = User::factory()->create();
-        $this->user->assignRole(Roles::SuperAdmin);
     }
 
     /** @test */
