+++ tests/Traits/InteractsWithPermissions.php
@@ -1,0 +1,53 @@
+<?php
+/*
+ * SSH Access Manager - SSH keys management solution.
+ *
+ * Copyright (c) 2017 - 2021 by Paco Orozco <paco@pacoorozco.info>
+ *
+ *  This file is part of some open source application.
+ *
+ *  Licensed under GNU General Public License 3.0.
+ *  Some rights reserved. See LICENSE, AUTHORS.
+ *
+ *  @author      Paco Orozco <paco@pacoorozco.info>
+ *  @copyright   2017 - 2021 Paco Orozco
+ *  @license     GPL-3.0 <http://spdx.org/licenses/GPL-3.0>
+ *  @link        https://github.com/pacoorozco/ssham
+ */
+
+namespace Tests\Traits;
+
+use Database\Seeders\PermissionsTableSeeder;
+use Database\Seeders\RolesTableSeeder;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Gate;
+
+trait InteractsWithPermissions
+{
+    public function disablePermissionsCheck(): void
+    {
+        Gate::before(function () {
+            return true;
+        });
+    }
+
+    public function enablePermissionsCheck(): void
+    {
+        $this->loadRolesAndPermissions();
+    }
+
+    public function loadRolesAndPermissions(): void
+    {
+        $this->seed(PermissionsTableSeeder::class);
+        $this->seed(RolesTableSeeder::class);
+    }
+
+    protected function expectAuthorizationException(): static
+    {
+        $this->expectException(AuthorizationException::class);
+
+        return $this;
+    }
+}
+
+
